pyproject.toml
# pyproject.toml
[project]
name = "x-keyword-monitor"
version = "0.1.0"
dependencies = [
    "playwright",
    "python-dotenv",
]